+++ Arduino/eTextile_matrix_sensor_blob/eTextile_matrix_sensor_blob.h
@@ -10,15 +10,6 @@
 
 #include "blob.h"
 
-// Control pins to send values to the 8-BITs shift registers used on the E-256 PCB
-// shiftOut using SPI library : https://forum.arduino.cc/index.php?topic=52383.0
-// Arduino UNO - SPI PINS
-// DATA_PIN -> SPI:MOSI -> D11 // Pin connected to Data in (DS) of the firdt 74HC595 8-BIT shift register
-// CLOCK_PIN -> SPI:SCK -> D13 // Pin connected to clock pin (SH_CP) of the first 74HC595 8-BIT shift register
-// LATCH_PIN -> SPI:SS -> D10  // Pin connected to latch pin (ST_CP) of the first 74HC595 8-BIT shift register
-
-// Teensy - SPI PINS https://www.pjrc.com/teensy/td_libs_SPI.html
-
 #define  BUILTIN_LED          13
 #define  BUTTON_PIN           32  // Button on the eTextile Teensy shield
 #define  BAUD_RATE            230400
@@ -52,7 +43,7 @@ const int columnPins[COLS] = {
 
 volatile uint16_t minVals[ROW_FRAME] = {0};  // Array to store smallest values
 float32_t frameValues[ROW_FRAME] = {0};      // Array to store ofset input values
-uint16_t bilinIntOutput[NEW_FRAME] = {0};    // Bilinear interpolation Output buffer
+uint8_t bilinIntOutput[NEW_FRAME] = {0};     // Bilinear interpolation Output buffer
 uint8_t myPacket[ROW_FRAME] = {0};           // Array to store values to transmit
 
 #ifdef CORE_TEENSY
