+++ Arduino/eTextile_matrix_sensor_blob/framebuffer.cpp
@@ -13,6 +13,5 @@ char _fb_base;
 framebuffer_t *fb_framebuffer = (framebuffer_t *) &_fb_base;
 
 uint32_t fb_buffer_size() {
-  // return (MAIN_FB()->w * MAIN_FB()->h) * sizeof(uint8_t);
-  return (MAIN_FB()->w * MAIN_FB()->h) * sizeof(uint16_t); // We need uint16_t frame buffer to hold 10 bit ADC values!
+  return (MAIN_FB()->w * MAIN_FB()->h) * sizeof(uint8_t);
 }
